move presence code to a separate file
